Mongo worker template + many useful functions.
The template can be run by a process to serve hyperparameter space samples to mongo workers (once a few substitutions have been made into it).

The utils functions include several that help with easily defining a neural network architecture along with its hyperparameters that you want to optimize, including a function to create Keras models from a sample of your hyperparameter space. There is also the main function which prepares the template, starts the mongo workers, copies over needed functions and imports from your notebook, and starts the process to serve hyperparameter samples.

All of this is designed so that it should work well from a Jupyter notebook; you call `hyper_optimize` with the function you want optimized as well as a function to generate the hyperparameter space, and, ideally, the rest works from there: any imports your functions depend upon should be copied over, as well as the source code for the functions themselves. A temporary directory is created, and all of this is written into a file there so that it can be imported in the template file (this lets the serialization work properly). Necessary substitutions (such as the name of the function you're optimizing) are made into the template, then a process is started to run the template and serve hyperparameters as well as other processes to run your function with these samples. None of this blocks the notebook kernel, so you can continue working in your notebook while the hyperparameter optimization runs.